+++ lib/src/bones_api_entity_db_object_memory.dart
@@ -10,9 +10,9 @@ import 'bones_api_condition.dart';
 import 'bones_api_condition_encoder.dart';
 import 'bones_api_entity.dart';
 import 'bones_api_entity_db.dart';
+import 'bones_api_entity_db_object.dart';
 import 'bones_api_entity_reference.dart';
 import 'bones_api_extension.dart';
-import 'bones_api_initializable.dart';
 
 final _log = logging.Logger('DBObjectMemoryAdapter');
 
@@ -42,14 +42,15 @@ typedef DBMemoryObjectAdapter = DBObjectMemoryAdapter;
 /// A [SQLAdapter] that stores tables data in memory.
 ///
 /// Simulates a SQL Database adapter. Useful for tests.
-class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
+class DBObjectMemoryAdapter
+    extends DBObjectAdapter<DBObjectMemoryAdapterContext> {
   static bool _boot = false;
 
   static void boot() {
     if (_boot) return;
     _boot = true;
 
-    DBAdapter.registerAdapter([
+    DBObjectAdapter.registerAdapter([
       'object.memory',
       'obj.memory',
     ], DBObjectMemoryAdapter, _instantiate);
@@ -71,11 +72,12 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
   }
 
   DBObjectMemoryAdapter(
-      {bool generateTables = false,
-      Object? populateTables,
-      Object? populateSource,
-      EntityRepositoryProvider? parentRepositoryProvider,
-      String? workingPath})
+      {super.generateTables = false,
+      super.populateTables,
+      super.populateSource,
+      super.parentRepositoryProvider,
+      super.workingPath,
+      super.log})
       : super(
           'object.memory',
           1,
@@ -84,9 +86,6 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
               dialect: DBDialect('object'),
               transactions: true,
               transactionAbort: true),
-          populateSource: populateSource,
-          parentRepositoryProvider: parentRepositoryProvider,
-          workingPath: workingPath,
         ) {
     boot();
 
@@ -124,14 +123,6 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
     );
 
     return adapter;
-  }
-
-  @override
-  List<Initializable> initializeDependencies() {
-    var parentRepositoryProvider = this.parentRepositoryProvider;
-    return <Initializable>[
-      if (parentRepositoryProvider != null) parentRepositoryProvider
-    ];
   }
 
   @override
@@ -267,143 +258,14 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
     var fieldsTypes =
         entityFieldsTypes.map((key, value) => MapEntry(key, value.type));
 
-    var fieldsReferencedTables = _findFieldsReferencedTables(table,
-        entityHandler: entityHandler, entityFieldsTypes: entityFieldsTypes);
-
-    var relationshipTables = _findRelationshipTables(table);
-
-    fieldsTypes.removeWhere((key, value) =>
-        relationshipTables.any((r) => r.relationshipField == key));
-
     tableScheme = TableScheme(table,
         relationship: relationship != null,
         idFieldName: idFieldName,
-        fieldsTypes: fieldsTypes,
-        fieldsReferencedTables: fieldsReferencedTables,
-        relationshipTables: relationshipTables);
+        fieldsTypes: fieldsTypes);
 
     _log.info('$tableScheme');
 
     return tableScheme;
-  }
-
-  Map<String, TableFieldReference> _findFieldsReferencedTables(String table,
-      {Type? entityType,
-      EntityHandler<dynamic>? entityHandler,
-      Map<String, TypeInfo>? entityFieldsTypes,
-      bool onlyCollectionReferences = false}) {
-    entityHandler ??=
-        getEntityHandler(tableName: table, entityType: entityType);
-
-    entityFieldsTypes ??= entityHandler!.fieldsTypes();
-
-    var relationshipFields =
-        Map<String, TypeInfo>.fromEntries(entityFieldsTypes.entries.where((e) {
-      var fieldType = e.value;
-      if (fieldType.isCollection != onlyCollectionReferences) return false;
-
-      var entityType = _resolveEntityType(fieldType);
-      if (entityType == null || entityType.isBasicType) return false;
-
-      var typeEntityHandler =
-          entityHandler!.getEntityHandler(type: entityType.type);
-      return typeEntityHandler != null;
-    }));
-
-    var fieldsReferencedTables = relationshipFields.map((field, fieldType) {
-      var targetEntityType = _resolveEntityType(fieldType)!;
-      var targetEntityHandler =
-          getEntityHandler(entityType: targetEntityType.type);
-
-      String? targetName;
-      String? targetIdField;
-      Type? targetIdType;
-      if (targetEntityHandler != null) {
-        targetName = getEntityRepositoryByType(targetEntityHandler.type)?.name;
-        targetIdField = targetEntityHandler.idFieldName();
-        targetIdType =
-            targetEntityHandler.getFieldType(null, targetIdField)?.type;
-      }
-
-      targetName ??= targetEntityType.type.toString().toLowerCase();
-      targetIdField ??= 'id';
-      targetIdType ??= int;
-
-      var tableRef = TableFieldReference(
-          table, field, targetIdType, targetName, targetIdField, targetIdType);
-      return MapEntry(field, tableRef);
-    });
-    return fieldsReferencedTables;
-  }
-
-  TypeInfo? _resolveEntityType(TypeInfo fieldType) {
-    if (fieldType.isPrimitiveType) {
-      return null;
-    }
-
-    var entityType =
-        fieldType.isListEntityOrReference ? fieldType.arguments0! : fieldType;
-
-    if (!EntityHandler.isValidEntityType(entityType.type)) {
-      return null;
-    }
-
-    return entityType;
-  }
-
-  List<TableRelationshipReference> _findRelationshipTables(String table) {
-    var allRepositories = this.allRepositories();
-
-    var allTablesReferences = Map.fromEntries(allRepositories.values.map((r) {
-      var referencedTables = _findFieldsReferencedTables(r.name,
-          entityType: r.type,
-          entityHandler: r.entityHandler,
-          onlyCollectionReferences: true);
-      return MapEntry(r, referencedTables);
-    }).where((e) => e.value.isNotEmpty));
-
-    for (var refs in allTablesReferences.values) {
-      refs.removeWhere((field, refTable) => refTable.sourceTable != table);
-    }
-
-    allTablesReferences.removeWhere((repo, refs) => refs.isEmpty);
-
-    var relationships = allTablesReferences.entries.expand((e) {
-      var repo = e.key;
-      var refs = e.value;
-
-      return refs.values.map((ref) {
-        var sourceTable = ref.sourceTable;
-        var sourceField = ref.sourceField;
-
-        var targetTable = ref.targetTable;
-        var targetField = ref.targetField;
-
-        var sourceEntityHandler = repo.entityHandler;
-        var sourceFieldId = sourceEntityHandler.idFieldName();
-        var sourceFieldIdType =
-            sourceEntityHandler.getFieldType(null, sourceFieldId)?.type ?? int;
-
-        var relTable = '${sourceTable}__${sourceField}__rel';
-        var relSourceField = '${sourceTable}__$sourceFieldId';
-        var relTargetField = '${targetTable}__$targetField';
-
-        return TableRelationshipReference(
-          relTable,
-          sourceTable,
-          sourceFieldId,
-          sourceFieldIdType,
-          relSourceField,
-          targetTable,
-          ref.targetField,
-          ref.targetFieldType,
-          relTargetField,
-          relationshipField: sourceField,
-        );
-      });
-    }).toList();
-
-    return relationships;
   }
 
   @override
@@ -673,8 +535,7 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
     var idField = _getTableIDFieldName(table);
     var id = entry[idField];
 
-    _log.info(
-        '[transaction:${op.transactionId}] doInsert> INSERT INTO $table OBJECT `$id`');
+    _logTransactionOperation('doInsert', op, 'INSERT INTO $table OBJECT `$id`');
 
     if (id == null) {
       throw StateError("Can't determine object ID to store it: $fields");
@@ -683,6 +544,13 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
     map[id] = entry;
 
     return _finishOperation(op, id, preFinish);
+  }
+
+  void _logTransactionOperation(
+      String method, TransactionOperation op, Object query) {
+    if (log) {
+      _log.info('[transaction:${op.transactionId}] $method> $query');
+    }
   }
 
   @override
@@ -703,8 +571,7 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
       String entityName,
       Object id,
       PreFinishDBOperation? preFinish) {
-    _log.info(
-        '[transaction:${op.transactionId}] doUpdate> UPDATE INTO $table OBJECT `$id`');
+    _logTransactionOperation('doUpdate', op, 'UPDATE INTO $table OBJECT `$id`');
 
     var map = _getTableMap(table, true)!;
 
@@ -865,7 +732,7 @@ class DBObjectMemoryAdapter extends DBAdapter<DBObjectMemoryAdapterContext> {
 typedef DBMemoryObjectAdapterException = DBObjectMemoryAdapterException;
 
 /// Error thrown by [DBObjectMemoryAdapter] operations.
-class DBObjectMemoryAdapterException extends DBAdapterException {
+class DBObjectMemoryAdapterException extends DBObjectAdapterException {
   @override
   String get runtimeTypeNameSafe => 'DBObjectMemoryAdapterException';